requirements.txt
fastapi
uvicorn
numpy
torch
torchvision
matplotlib
pillow
python-multipart